pyproject.toml
[tool.poetry]
name = "telegram-vip-passport-backend"
version = "0.1.0"
package-mode = false
description = "Backend service for the Telegram VIP Passport ecosystem."
authors = ["Seylanediff Team <info@seylanediff.dev>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114"
uvicorn = {extras = ["standard"], version = "^0.23"}
sqlalchemy = "^2.0"
alembic = "^1.11"
psycopg2-binary = "^2.9"
pydantic-settings = "^2.1"
httpx = "^0.28"
aiogram = "^3.0"
redis = "^5.0"
asyncpg = "^0.29.0"
python-jose = {extras = ["cryptography"], version = "^3.3"}
aiosqlite = "^0.21.0"
greenlet = "^3.2.4"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
